created a track class/model and changed to using a map of playlists to store user playlists
