+++ models/track.js
@@ -1,0 +1,9 @@
+class Track  {
+  constructor(name, artist, uriReference) {
+    this.name = name;
+    this.artist = artist;
+    this.uriReference = uriReference;
+  }
+}
+
+module.exports = Track;
